Revise chewing fuzzer
1. Specify unique name for temp userdb. So we can run multiple fuzzers
   at the same time.
2. Reorganized as three fuzzers for different variants. So we can
   prioritize easily (default first, dynamic config last)
